services/backend/requirements.txt
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
google-generativeai==0.8.3
